+++ all_test.sh
@@ -5,4 +5,4 @@ pytest tests/
 pytest packages/oasa/tests/
 pytest packages/bkchem-app/tests/
 QT_QPA_PLATFORM=cocoa
-pytest packages/bkchem-qt.app/tests/
+pytest packages/bkchem-qt.app/tests/ -s