+++ packages/bkchem-qt.app/bkchem_qt/modes/bracket_mode.py
@@ -143,8 +143,9 @@ class BracketMode(bkchem_qt.modes.base_mode.BaseMode):
 
 
 #============================================
-def _make_rect(p1: PySide6.QtCore.QPointF,
-			   p2: PySide6.QtCore.QPointF) -> PySide6.QtCore.QRectF:
+def _make_rect(
+		p1: PySide6.QtCore.QPointF,
+		p2: PySide6.QtCore.QPointF) -> PySide6.QtCore.QRectF:
 	"""Build a QRectF from two corner points.
 
 	Args: